+++ akd_client/src/lib.rs
@@ -101,6 +101,8 @@ pub mod verify;
 pub(crate) mod ecvrf;
 pub(crate) mod hash;
 pub(crate) mod utils;
+#[cfg(any(test, feature = "converters"))]
+pub(crate) mod converters;
 
 /// The arity of the tree. Should EXACTLY match the ARITY within
 /// the AKD crate (i.e. akd::ARITY)
